% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_labels}
\alias{add_labels}
\alias{add_labels.collection}
\title{Adds labels}
\usage{
add_labels(obj, labels, ids_from, labels_from, ...)

add_labels(obj, labels, ids_from, labels_from, ...)
}
\arguments{
\item{obj}{A spectra collection}

\item{labels}{A dataframe or tibble associating sample ids to labels}

\item{ids_from}{Name of the id column}

\item{labels_from}{Name of the label column}

\item{...}{further arguments passed to or from other methods(not
currenctly used).}
}
\value{
An updated version of \code{collection}.
}
\description{
Add labels to an existing collection by providing a
dat frame or tibble associating sample ids to labels
}
\details{
Pre-existing labels will be overwritten.
}
\examples{
library(tidySpectR)

newlabs <- tibble(names = pull_ids(fa_nmr),
                  conditions = c("organic", "organic", "organic", "organic", "organic",
                                 "conventional", "conventional", "conventional", "conventional", "conventional"))

add_labels(fa_nmr, newlabs, ids_from= "names", labels_from= "conditions")
}
