% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aibin_cpp}
\alias{aibin_cpp}
\title{Adaptive Intelligent Binning Algorithm}
\usage{
aibin_cpp(spectra, noise, r)
}
\arguments{
\item{spectra, }{noise A mtrix containing spectra or nosie regions.
Cols 0-2  are : bins, bin_start, bin_end, then come samples (col-wise)}

\item{R}{resolution (0>R>=1)}
}
\value{
vector of dbl containing the split position in the orignal axis unit.
}
\description{
Entry point from R for the C++ implementation
}
\keyword{internal}
