% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{average_spectrum}
\alias{average_spectrum}
\alias{average_spectrum.collection}
\title{Calculate the average spectrum}
\usage{
average_spectrum(x, ...)

\method{average_spectrum}{collection}(x, group = "all", ...)
}
\arguments{
\item{x}{A\code{collection} object}

\item{...}{further arguments passed to or from other methods(not
currenctly used).}

\item{group}{How to perform grouping. Either \code{all}or \code{labels}}
}
\value{
An updated version of \code{collection}.
}
\description{
Averages spectra
}
\examples{
library(tidySpectR)

average_spectrum(fa_nmr)

average_spectrum(fa_nmr, group = 'labels')
}
