% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket_aibin.R
\name{bin_eval_test}
\alias{bin_eval_test}
\title{Bin Evaluation Criterion}
\usage{
bin_eval_test(vbmax, vb, vb1max, vb2max, vnoise)
}
\arguments{
\item{vbmax}{sum of maximums of left and right bin values}

\item{vb}{original bin value}

\item{vb1max}{max bin value of left bin}

\item{vb2max}{max bin value of right bin}

\item{vnoise}{minimal bin value}
}
\value{
TRUE if splitting improves bin value, FALSE otherwise
}
\description{
Evaluates whether a bin can be split or not
}
\keyword{internal}
