% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket_aibin.R
\name{bin_value_inner_product}
\alias{bin_value_inner_product}
\title{Computes the inner product of the bin value equation
for a dataset with a single sample}
\usage{
bin_value_inner_product(dat, id, R)
}
\arguments{
\item{dat}{A tibble or dat frame with columns
\describe{
\item{bin_start}{left border of the bin}
\item{bin_end}{right border of the bin}
\item{values}{bin intensity}
}}

\item{id}{will be ignored}

\item{R}{resolution}
}
\value{
a tibble with column product containing
the inner product of the bin value equation for 1 sample
}
\description{
Computes the inner product of the bin value equation
for a dataset with a single sample
}
\keyword{internal}
