% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket_aibin.R
\name{bucket_aibin}
\alias{bucket_aibin}
\title{Applies Adaptive Inteligent Binning on a collection`}
\usage{
bucket_aibin(x, R, noise_region, cores = 1, ...)

bucket_aibin(x, R, noise_region, cores = 1, ...)
}
\arguments{
\item{x}{A\code{collection} object to bucket}

\item{R}{resolution value, strictly positive and typically in the interval
0 > R >= 1. A lower R value will results in more bins being created.}

\item{noise_region}{A\code{collection} object containing a noise_region}

\item{cores}{Number of cores to allocate to the process for multprocessing}

\item{...}{further arguments passed to or from other methods(not
currenctly used).}
}
\value{
An updated version of x
}
\description{
Adaptive Inteligent Binning (aibin) recursively finds bin edges
in existing bins by maximizing the information contained in each
newly created bin.
}
\details{
AI binning will first consider the entire spectra as a single bin.
It will then consider every position in the spectra as a potential
splitting candidate and calulate a quality value for each splitting candidate.
If the best splitting candidates fulfills quality requirements, the bin
will be split and the process will be repeated on the newly created bins.

In order to limit splitting in local minima of noisy regions a minimal quality
value is computed from the user-provided noise-region.
}
\examples{
\dontrun{
library(tidySpectR)

# First we normalize the dataset and extract noise and sepctra regions
normalized <- fa_nmr \%>\% 
                  normalize_internalStandard(3.58, 3.64)
spectra <- normalized \%>\% extract(0, 7.2)
noise <- normalized \%>\% mask(0, 7.2, overlaps = 'remove')

# Sit back and relax
bucketted <- bucket_aibin(spectra, 0.2, noise, cores  = 2)
}
}
\references{
De Meyer T, Sinnaeve D, Van Gasse B, Tsiporkova E, Rietzschel ER,
De Buyzere ML, Gillebert TC, Bekaert S, Martins JC, Van Criekinge W.
NMR-based characterization of metabolic alterations in hypertension
using an adaptive, intelligent binning algorithm. Anal Chem. 2008
May 15;80(10):3783-90. doi: 10.1021/ac7025964.
\url{https://pubmed.ncbi.nlm.nih.gov/18419139/}
}
