% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucketting.R
\name{bucket_from_breaks}
\alias{bucket_from_breaks}
\alias{bucket_from_breaks.collection}
\title{Bucket a collection form a list of breaks (bucket limits).
For internal use}
\usage{
bucket_from_breaks(...)

\method{bucket_from_breaks}{collection}(obj, breaks, ...)
}
\arguments{
\item{...}{further arguments passed to or from other methods(not
currenctly used).}

\item{obj}{A\code{collection} object}

\item{breaks}{A vector of breaking values}
}
\value{
An updated version of \code{collection}.
}
\description{
Bucket a collection form a list of breaks (bucket limits).
For internal use
}
\details{
If omitted the spectra limits will be added o the breaks

Buckets will include the bin overlapping on the lower limit and exclude the bin
overlapping on the higher limit.
}
