% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket_optimized.R
\name{bucket_optimized}
\alias{bucket_optimized}
\alias{bucket_optimized.collection}
\title{Performs optimized bucketting}
\usage{
bucket_optimized(x, ...)

\method{bucket_optimized}{collection}(x, initial_width, slackness, skip = FALSE, ...)
}
\arguments{
\item{x}{A\code{collection} object}

\item{...}{further arguments passed to or from other methods(not
currenctly used).}

\item{initial_width}{Initial bin width to optimize}

\item{slackness}{fraction of initial width that defines
how far the boundary can move while searching for a local minimum}

\item{skip}{Skip the creation of of processor step. If TRUE, this step will not be added to
the list of processing steps. Typically reserved for nested function calls.}
}
\value{
An updated version of \code{collection}.
}
\description{
Bucket a collection of spectra using the Optimized
Bucketting Algorithm
}
\details{
The Optimized Bucketting Algorithm optimizes bucket size by
searching local minima in the average spectrum.
}
\examples{
library(tidySpectR)

bucket_optimized(fa_nmr, initial_width = 0.1, slackness = 0.2)
}
\references{
Sousa SAA, Magalhaes A, Ferreira MMC. Optimized bucketing for NMR spectra:
Three case studies. Chemometr Intell Lab. 2013;122:93–102.
https://doi.org/10.1016/j.chemolab.2013.01.006
\url{https://doi.org/10.1016/j.chemolab.2013.01.006}
}
