% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{collection.default}
\alias{collection.default}
\title{rdname collection}
\usage{
\method{collection}{default}(...)
}
\arguments{
\item{...}{further arguments passed to or from other methods(not
currenctly used).}
}
\value{
An object of class \code{collection} with sub-objects:
\item{data}{A tibble that contains the spectral data}
\item{normalized}{Information about the normalization of the dataset}
\item{bucketted}{Information about the bucketting of the dataset}
\item{labels}{Information about the labelling of the dataset}
}
\description{
rdname collection
}
\details{
The collection is empty on creation, use \code{add_spectrum} to
populate it.
}
