% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{data2wide}
\alias{data2wide}
\title{Pivot_wider on collection$data}
\usage{
data2wide(x, ids = "id", ...)
}
\arguments{
\item{x}{data tibble}

\item{...}{not currently used}
}
\description{
Pivot_wider on collection$data
}
\keyword{internal}
