% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{divide_bin}
\alias{divide_bin}
\title{Find a new division splitting position for the current bin}
\usage{
divide_bin(bin, r, noise, left, right)
}
\arguments{
\item{bin}{A matrix containing the whole spectral range.
Cols 0-2  are : bins, bin_start, bin_end, then come samples (col-wise)}

\item{r}{resolution}

\item{noise}{noise value}

\item{left, right}{index of bin_end for the current bin
(C++ index 0 to length-1).}
}
\value{
A dbl, the new optimal splitting position or
NAN if not splitting is possible
}
\description{
Find a new division splitting position for the current bin
}
\keyword{internal}
