% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{export_processor}
\alias{export_processor}
\alias{export_processor.collection}
\title{Exports a processor object}
\usage{
export_processor(...)

\method{export_processor}{collection}(x, ...)
}
\arguments{
\item{...}{further arguments passed to or from other methods(not
currenctly used).}

\item{x}{A \code{collection} object}
}
\value{
A \code{processing_template} object
}
\description{
Exports a processor object
}
\examples{
library(tidySpectR)

fa_nmr \%>\% 
  mask(from = 5, to = Inf) \%>\%
  bucket_uniform(N = 10) \%>\%
  export_processor() \%>\%
  tidy()
}
