% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_meas2.R
\name{f_meas2_vec}
\alias{f_meas2_vec}
\title{Simple wrapper for yardstick::f_meas with beta = 0.5 to use in tuning operations}
\usage{
f_meas2_vec(truth, estimate, estimator = NULL, na_rm = TRUE, ...)
}
\description{
Simple wrapper for yardstick::f_meas with beta = 0.5 to use in tuning operations
}
