% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fa_nmr}
\alias{fa_nmr}
\title{H1-NMR spectra of Egg yolk fatty-acids}
\format{
A collection object
\describe{
\item{data}{Tibble containing the spectral data}
\item{normalized}{Normalization method flag}
\item{bucketted}{Bucketting method flag}
\item{labels}{Tibble associating each sample to a condition}
}
}
\usage{
fa_nmr
}
\description{
A dataset containing H1-NMR fatty-acid spectra for 10 egg yolk samples, binned to 0.0004 ppm.
}
\keyword{datasets}
