% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoise_snr.R
\name{filter_noise}
\alias{filter_noise}
\alias{filter_noise.collection}
\title{Remove noise bins}
\usage{
filter_noise(x, ...)

\method{filter_noise}{collection}(x, noise_region, snr, ...)
}
\arguments{
\item{x}{A\code{collection} object to bucket}

\item{...}{further arguments passed to or from other methods(not
currenctly used).}

\item{noise_region}{A\code{collection} object containing a noise_region}

\item{snr}{A dbl, signal to noise ratio}
}
\value{
An updated version of x
}
\description{
Average maximal intensity is calculated form the noise region
and each bin in the spectra. Bins whose average maximal intensity
are not above the signal to noise ratio are discarded.
}
\examples{
library(tidySpectR)

# First we normalize the dataset and extract noise and sepctra regions
normalized <- fa_nmr \%>\% 
                  normalize_internalStandard(3.58, 3.64)
spectra <- normalized \%>\% extract(0, 7.2)
noise <- normalized \%>\% mask(0, 7.2, overlaps = 'remove')

filtered <- filter_noise(spectra, noise, 3)
}
