% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket_optimized.R
\name{find_local_min}
\alias{find_local_min}
\title{Find local minimum between two boundaries}
\usage{
find_local_min(i, data, N, s)
}
\arguments{
\item{i}{A counter}

\item{data}{A dataframe with columns values}

\item{N}{number of point per bin (initial)}

\item{s}{slackness}
}
\description{
For internal use only.
}
