% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket_aibin.R
\name{find_vnoise}
\alias{find_vnoise}
\title{Find the max bin value in the noise region}
\usage{
find_vnoise(noise_region, R)
}
\arguments{
\item{noise_region}{A \code{collection} object containing the noise region}

\item{R}{resolution}
}
\value{
A dbl, maximum bin value in the noise region
}
\description{
Find the max bin value in the noise region
}
\keyword{internal}
