% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_extract.R
\name{mask}
\alias{mask}
\alias{mask.collection}
\title{Masks a spectral region}
\usage{
mask(x, ...)

\method{mask}{collection}(x, from, to, overlaps = "keep", skip = FALSE, ...)
}
\arguments{
\item{x}{A\code{collection} object}

\item{...}{further arguments passed to or from other methods(not
currenctly used).}

\item{from, to}{Coordinates of the region to mask.
Use -Inf or Inf to shorten the spectra.}

\item{overlaps}{What to do with the bins ovelapping the edge of the mask.
Use either \code{keep} of \code{remove}.}

\item{skip}{Skip the creation of of processor step. If TRUE, this step will not be added to
the list of processing steps. Typically reserved for nested function calls.}
}
\value{
An updated version of \code{collection}.
}
\description{
\code{mask} discards a spectral region
}
\details{
The upper limit in not included: [min, max).
}
\examples{
library(tidySpectR)

# Trimming one side
mask(fa_nmr, from = Inf, to = 7.2)

# Masking region
mask(fa_nmr, from = 3.5, to = 3, overlaps = 'remove')
}
