% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{median_spec}
\alias{median_spec}
\title{Median spectrum}
\usage{
median_spec(x, ...)
}
\arguments{
\item{x}{data tibble (long form)}

\item{...}{not currently used}
}
\description{
Median spectrum
}
\keyword{internal}
