% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{median_spectrum}
\alias{median_spectrum}
\alias{median_spectrum.collection}
\title{Calculate the average spectrum}
\usage{
median_spectrum(x, ...)

\method{median_spectrum}{collection}(x, group = "all", ...)
}
\arguments{
\item{x}{A\code{collection} object}

\item{...}{further arguments passed to or from other methods(not
currenctly used).}

\item{group}{How to perform grouping. Either \code{all}or \code{labels}}
}
\value{
An updated version of \code{collection}.
}
\description{
Calculates median of spectra
}
\examples{
library(tidySpectR)

median_spectrum(fa_nmr)

median_spectrum(fa_nmr, group = 'labels')
}
