% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_template.R
\name{new_step}
\alias{new_step}
\alias{new_step.processing_template}
\title{Add a new \code{processing_step}}
\usage{
new_step(...)

\method{new_step}{processing_template}(x, fun, arglist, name = "none", id = NULL, ...)
}
\arguments{
\item{...}{further arguments passed to or from other methods(not
currenctly used).}

\item{x}{A \code{processing_template} object}

\item{fun}{either a function or a non-empty character string naming the function to be called.}

\item{arglist}{a list of arguments to the function call. The names attribute of args gives the argument names.}

\item{name}{A string description of the step}

\item{id}{An id associated with the step. Will be converted to a unique id.}
}
\value{
A list-like object storing processing steps.
}
\description{
Add a new \code{processing_step}
}
\examples{
library(tidySpectR)

template <- processing_template() \%>\%
            new_step(mask, list(from = 5, to = Inf), name = "mask") \%>\%
            new_step(bucket_uniform, list(N = 10), name = "bucket")
template
}
