% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nmr_spectra}
\alias{nmr_spectra}
\title{H1-NMR spectra of Egg yolk fatty-acids}
\format{
A tibble with 2621440 rows and 3 variables:
\describe{
\item{ID}{Unique sample identifier}
\item{bins}{Ppm value (x-axis)}
\item{values}{Intensity value}
}
}
\usage{
nmr_spectra
}
\description{
A dataset containing H1-NMR fatty-acid spectra for 10 egg yolk samples, binned to 0.0004 ppm.
}
\keyword{datasets}
