% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_factor.R
\name{normalize_factor}
\alias{normalize_factor}
\alias{normalize_factor.collection}
\title{Normalizes spectra to a given factor}
\usage{
normalize_factor(x, ...)

\method{normalize_factor}{collection}(x, factors, ...)
}
\arguments{
\item{x}{A\code{collection} object}

\item{...}{further arguments passed to or from other methods(not
currenctly used).}

\item{factors}{A tibble with columns \code{id} and \code{factor}}
}
\value{
An updated version of \code{collection}.
}
\description{
Intensities will be multiplied by the given factor.
}
\details{
Will mutliply the spectra by the factors given
as arguments.

Missing ids will not be normalized

This method should be used to bucket new data based on the results from
a previous dataset.
}
\examples{
library(tidySpectR)
library(tibble)

norm <- tibble(id = pull_ids(fa_nmr), factors = runif(10))

normalize_factor(fa_nmr, norm)
}
