% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_internalStandard.R
\name{normalize_internalStandard}
\alias{normalize_internalStandard}
\alias{normalize_internalStandard.collection}
\title{Normalizes spectra to an internal standard}
\usage{
normalize_internalStandard(x, ...)

\method{normalize_internalStandard}{collection}(x, from, to, skip = FALSE, ...)
}
\arguments{
\item{x}{A\code{collection} object}

\item{...}{arguments to be passed to \code{extract}}

\item{from, to}{Coordinates of the region to use as standard.}

\item{skip}{Skip the creation of of processor step. If TRUE, this step will not be added to
the list of processing steps. Typically reserved for nested function calls.}
}
\value{
An updated version of \code{collection}.
}
\description{
The integral of the spectrum between the given limits
will be used as a normalization factor. This assumes that
the selected peak intensity in independent of the experimental conditions.
}
\examples{
library(tidySpectR)

normalize_internalStandard(fa_nmr, from = 3, to = 3.5)
}
