% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_probQuotient.R
\name{normalize_probQuotient}
\alias{normalize_probQuotient}
\alias{normalize_ProbQuotient.collection}
\alias{normalize_probQuotient.collection}
\title{Probabilistic Quotient normalization}
\usage{
normalize_probQuotient(x, ...)

\method{normalize_probQuotient}{collection}(x, ...)
}
\arguments{
\item{x}{A\code{collection} object to normalize}

\item{...}{further arguments passed to or from other methods(not
currently used).}
}
\description{
Probabilistic Quoteint Normalization estimates for each spectra the
the most likely quotient derived form the ratio of the signal distribution
and that oif a reference (median) spectrum.
}
\examples{
library(tidySpectR)

normalize_probQuotient(fa_nmr)
}
\references{
Dieterle, F., Ross, A., Schlotterbeck, G., & Senn, H. (2006). Probabilistic
Quotient Normalization as Robust Method to Account for Dilution of Complex
Biological Mixtures. Application in1H NMR Metabonomics. Analytical Chemistry,
78(13), 4281–4290. doi:10.1021/ac051632c
\url{https://pubmed.ncbi.nlm.nih.gov/16808434/}
}
