% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_totalSpectrum.R
\name{normalize_totalSpectrum}
\alias{normalize_totalSpectrum}
\alias{normalize_totalSpectrum.collection}
\title{Normalizes spectra to their integrals}
\usage{
normalize_totalSpectrum(...)

\method{normalize_totalSpectrum}{collection}(obj, ...)
}
\arguments{
\item{...}{further arguments passed to or from other methods(not
currenctly used).}

\item{obj}{A\code{collection} object}
}
\value{
An updated version of \code{collection}.
}
\description{
Sums up the signal over the whole spectrum and use it as a
normalization factor. This assumes that the total signal
varies proportionally to the sample concentration
}
\examples{
library(tidySpectR)

normalize_totalSpectrum(fa_nmr)
}
