% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_template.R
\name{print.processing_template}
\alias{print.processing_template}
\title{Print processing_template}
\usage{
\method{print}{processing_template}(x, ...)
}
\arguments{
\item{x}{A \code{processing_step}object}

\item{...}{further arguments passed to or from other methods(not
currenctly used).}
}
\value{
The original object (invisibly)
}
\description{
Print processing_template
}
