% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_step.R
\name{process}
\alias{process}
\alias{process.processing_step}
\title{Applies a processing step on a \code{collection} object}
\usage{
process(x, ...)

\method{process}{processing_step}(x, collection, ...)
}
\arguments{
\item{x}{A \code{processing_step} object to apply}

\item{...}{further arguments passed to or from other methods(not
currenctly used).}

\item{collection}{A \code{collection} object to process}
}
\value{
The retrun value of the processing step call
}
\description{
Applies a processing step on a \code{collection} object
}
\examples{
library(tidySpectR)

# Creating a function call to `mask`
# Note that the `collection` argument is omitted
fun_call <- as.list(match.call(mask, call("mask", from = 5, to = Inf)))

stp <- processing_step(fun_call, "masking", "masking_01")
processed <- process(stp, fa_nmr)
processed
}
