% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_step.R
\name{processing_step}
\alias{processing_step}
\alias{processing_step.default}
\alias{tidy.processing_step}
\title{Stores specifications of a \code{collection} processing step}
\usage{
processing_step(...)

\method{processing_step}{default}(fun, arglist, name = "none", id = "none", ...)

\method{tidy}{processing_step}(x, ...)
}
\arguments{
\item{...}{further arguments passed to or from other methods(not
currenctly used).}

\item{fun}{either a function or a non-empty character string naming the function to be called.}

\item{arglist}{a list of arguments to the function call. The names attribute of args gives the argument names.}

\item{name}{A string description of the step}

\item{id}{An id associated with the step. Will be converted to a unique id.}

\item{x}{A \code{processing_step}object}
}
\value{
An object of class \code{processing_step}

A tibble representation of x
}
\description{
Processing steps specification are stored in a step
object. This will be typically handled internally.
}
\details{
A \code{processing_step} stores the function call and associated
arguments to reproduce a processing step at a later time
point a a different collection object.
}
\examples{
library(tidySpectR)

# Creating a masking step
stp <- 
   processing_step(
     fun = mask, 
     arglist = list(from = 5, to = Inf),
     type = "masking", 
     name = "masking")

stp
library(tidySpectR)

# Creating a masking step
stp <- 
   processing_step(
     fun = mask, 
     arglist = list(from = 5, to = Inf),
     type = "masking", 
     name = "masking")

tidy(stp)
}
