% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_template.R
\name{processing_template}
\alias{processing_template}
\alias{processing_template.default}
\title{Stores a template for processing of \code{collection} objects}
\usage{
processing_template(...)

\method{processing_template}{default}(...)
}
\arguments{
\item{...}{further arguments passed to or from other methods(not
currenctly used).}
}
\value{
A list-like object storing processing steps.
}
\description{
#' @param ... further arguments passed to or from other methods(not
currenctly used).
}
