% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pull_breaks}
\alias{pull_breaks}
\alias{pull_breaks.collection}
\title{Pull the breaking points (or bin limits) of a spectra collection}
\usage{
pull_breaks(obj, ...)

\method{pull_breaks}{collection}(obj, ...)
}
\arguments{
\item{obj}{A spectra collection}

\item{...}{further arguments passed to or from other methods(not
currenctly used).}
}
\value{
A vector of dbl.
}
\description{
For a collection, retrieve the bin limits, including the edges
(min and max limits of the spectra).
}
