% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pull_limits}
\alias{pull_limits}
\alias{pull_limits.collection}
\title{Pull the limits of a spectra collection}
\usage{
pull_limits(x, ...)

\method{pull_limits}{collection}(x, ...)
}
\arguments{
\item{x}{A spectra collection}

\item{...}{further arguments passed to or from other methods(not
currenctly used).}
}
\value{
A vector of int.
}
\description{
For a collection, the upper and lower limits of the x-axis
}
\examples{
library(tidySpectR)
pull_limits(fa_nmr)
}
