% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{pull_numbin}
\alias{pull_numbin}
\alias{pull_numbin.collection}
\title{Pull the bin (or bucket) number of a spectra collection}
\usage{
pull_numbin(obj, ...)

\method{pull_numbin}{collection}(obj, ...)
}
\arguments{
\item{obj}{A spectra collection}

\item{...}{further arguments passed to or from other methods(not
currenctly used).}
}
\value{
A tibble.
}
\description{
For a collection, retrieve the number of bins.
}
