% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_step.R
\name{rand_id}
\alias{rand_id}
\title{Generate random id}
\usage{
rand_id(id, ...)
}
\arguments{
\item{id}{A base string to expand with random digits}
}
\description{
Generate random id
}
\keyword{internal}
