% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{recc_split}
\alias{recc_split}
\title{Reccursively split the bin}
\usage{
recc_split(bin, r, noise, left, right, indices)
}
\arguments{
\item{bin}{A matrix containing the whole spectral range.
Cols 0-2  are : bins, bin_start, bin_end, then come samples (col-wise)}

\item{r}{resolution}

\item{noise}{noise value}

\item{left, right}{index of bin_end for the current bin
(C++ index 0 to length-1).}

\item{indices}{storage of splitting indices, should be initializied with \code{right}.}
}
\value{
A vector of splitting indices
}
\description{
Reccursively split the bin
}
\keyword{internal}
