% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket_aibin.R
\name{recursive_split}
\alias{recursive_split}
\title{Recursively split the spectra with the adaptive inteligent binning algorithm}
\usage{
recursive_split(current_bin, R, vnoise, breaks = NULL)
}
\arguments{
\item{current_bin}{A \code{collection} object containing the current bin}

\item{R}{resolution}

\item{vnoise}{Maximal bin value for the noise region.}

\item{breaks}{A vector of double that will be populated with the
binning positions}
}
\value{
A vector of dbl contianing the split positions
}
\description{
Recursively split the spectra with the adaptive inteligent binning algorithm
}
\keyword{internal}
