% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_pareto.R
\name{step_pareto}
\alias{step_pareto}
\alias{tidy.step_pareto}
\title{Applies Pareto scaling on numeric data}
\usage{
step_pareto(
  recipe,
  ...,
  role = NA,
  trained = FALSE,
  means = NULL,
  sdroots = NULL,
  na_rm = TRUE,
  skip = FALSE,
  id = rand_id("pareto")
)

\method{tidy}{step_pareto}(x, ...)
}
\arguments{
\item{...}{One or more selector functions to choose which
variables are affected by the step. See \code{\link[=selections]{selections()}}
for more details. For the \code{tidy} method, these are not
currently used.}

\item{role}{Not used by this step since no new variables are
created.}

\item{means}{A named numeric vector of means. This is
\code{NULL} until computed by \code{\link[=prep.recipe]{prep.recipe()}}.}

\item{sdroots}{A named numeric vector of standard deviation square roots. This
is \code{NULL} until computed by \code{\link[=prep.recipe]{prep.recipe()}}.}

\item{na_rm}{A logical value indicating whether \code{NA}
values should be removed when computing the standard deviation and mean.}

\item{x}{A \code{step_pareto} object.}
}
\value{
An updated version of \code{recipe} with the new step
added to the sequence of existing steps (if any). For the
\code{tidy} method, a tibble with columns \code{terms} (the
selectors or variables selected), \code{value} (the
standard deviations and means), and \code{statistic} for the type of value.
}
\description{
\code{step_pareto} creates a \emph{specification} of a recipe
step that will perform Pareto scaling on the columns.
}
\details{
Pareto scaling is a variant of autoscaling whereby the data is scaled
by the square root of its standard deviation. \code{step_pareto} estimates the standard deviations
and means from the data used in the \code{training} argument of
\code{prep.recipe}. \code{bake.recipe} then applies the scaling to new data sets using
these estimates.
}
\examples{
# requires the recipes package
pareto <- 
  recipe(Species ~. , iris) \%>\%
  step_pareto(all_predictors())
}
\references{
van den Berg, R. A., Hoefsloot, H. C., Westerhuis, J. A., Smilde, A. K., &
van der Werf, M. J. (2006). Centering, scaling, and transformations:
improving the biological information content of metabolomics data.
BMC genomics, 7, 142. https://doi.org/10.1186/1471-2164-7-142
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC1534033/}
}
\concept{normalization_methods}
\concept{preprocessing}
\keyword{datagen}
