% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_x_vast.R
\name{step_xvast}
\alias{step_xvast}
\title{Applies x-VAST scaling on numeric data}
\usage{
step_xvast(
  recipe,
  ...,
  scaling = "autoscale",
  role = NA,
  trained = FALSE,
  outcome = NULL,
  means = NULL,
  sds = NULL,
  cvs = NULL,
  na_rm = TRUE,
  skip = FALSE,
  id = rand_id("xvast")
)
}
\arguments{
\item{...}{One or more selector functions to choose which
variables are affected by the step. See \code{\link[=selections]{selections()}}
for more details. For the \code{tidy} method, these are not
currently used.}

\item{scaling}{Either \code{autoscale} or \code{pareto}. Controls the scaling method.
See notes below.}

\item{role}{Not used by this step since no new variables are
created.}

\item{outcome}{When a single outcome is available, character
string or call to \code{\link[dplyr:vars]{dplyr::vars()}} can be used to specify a single outcome
variable.}

\item{means}{A named numeric vector of means. This
is \code{NULL} until computed by \code{\link[=prep.recipe]{prep.recipe()}}.}

\item{sds}{A named numeric vector of stadard deviations. This
is \code{NULL} until computed by \code{\link[=prep.recipe]{prep.recipe()}}.}

\item{cvs}{A named numeric vector of variation coeficients. This
is \code{NULL} until computed by \code{\link[=prep.recipe]{prep.recipe()}}.}

\item{na_rm}{A logical value indicating whether \code{NA}
values should be removed when computing the standard deviation and mean.}
}
\value{
An updated version of \code{recipe} with the new step
added to the sequence of existing steps (if any). For the
\code{tidy} method, a tibble with columns \code{terms} (the
selectors or variables selected), \code{value} (the
standard deviations and means), and \code{statistic} for the type of value.
}
\description{
\code{step_svast} creates a \emph{specification} of a recipe
step that will perform x-VAST scaling on the columns
}
\details{
supervised maximum Variable Stability (x-VAST) scaling preforms centering and scaling followed
by a weighting of each variable by the maximum of the class-specific variation coeficients.

The argument \code{scaling} controls which scaling method should be used before
variable weighting. \code{autoscale} will perform mean-centering and standard deviation
scaling while \code{pareto} will scale by the square-root of the standard deviation.
}
\examples{
# requires the recipes package
autoscale_xvast <- 
  recipe(Species ~. , iris) \%>\%
  step_vast(all_predictors(), scaling = 'autoscale', outcome = 'Species')

pareto_svast <- 
  recipe(Species ~. , iris) \%>\%
  step_xast(all_predictors(), scaling = 'pareto', outcome = 'Species')
}
\references{
Yang, J., Zhao, X., Lu, X., Lin, X., & Xu, G. (2015). A data preprocessing
strategy for metabolomics to reduce the mask effect in data analysis.
Frontiers in molecular biosciences, 2, 4. https://doi.org/10.3389/fmolb.2015.00004
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4428451/}
}
\concept{normalization_methods}
\concept{preprocessing}
\keyword{datagen}
