% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subsample}
\alias{subsample}
\alias{subsample.collection}
\title{Sub-sample a spectra collection}
\usage{
subsample(x, ...)

\method{subsample}{collection}(x, size = 2, balance_labels = FALSE, ids = NULL, ...)
}
\arguments{
\item{x}{A\code{collection} object}

\item{...}{further arguments passed to or from other methods(not
currenctly used).}

\item{size}{Size of the sub-sampling}

\item{balance_labels}{A boolean, wether to pick spectra equally between labels.
If the subsampling size is below the number of labels, not all labels will be chosen from.}

\item{ids}{A vector of ids to extract, this will override the other parameters.}
}
\value{
a subsample of x
}
\description{
Pick a subsample form a spectra collection
}
\examples{
library(tidySpectR)
library(dplyr)

# Randomly sample
subsample(fa_nmr, size = 3)

# Sample one spectra for each labels
subsample(fa_nmr, size= 2, balance_labels = TRUE)

# Subsample specific IDs
subsample(fa_nmr, ids = c("20199305928", "20199306281", "20199504645"))

# Sample alll spectra of specific labels
organic <- fa_nmr$labels \%>\% 
           filter(label == 'organic') \%>\%
           pull(id)
subsample(fa_nmr, ids = organic)

# Chain subsamplings to randomly select a spectra with a specific label
subsample(fa_nmr, ids = organic) \%>\%
  subsample(size=1)
}
