% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.collection}
\alias{tidy.collection}
\title{Tidy a spectra collection}
\usage{
\method{tidy}{collection}(x, ...)
}
\arguments{
\item{x}{A \code{collection} object.}

\item{...}{further arguments passed to or from other methods(not
currenctly used).}
}
\value{
A tibble with a variable number columns
depending on the binning of the data:
\item{id}{Unique sample identifier}
\item{label}{Label for the sample}
\item{bins}{Bin centers}
}
\description{
\code{tidy} will return a dataframe containing the spectral
information in its current state.
}
\examples{
library(tidySpectR)

tidy(fa_nmr)
}
