% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_step.R
\name{tidy.processing_step}
\alias{tidy.processing_step}
\title{Tidy a \code{processing_step}}
\usage{
\method{tidy}{processing_step}(x, ...)
}
\arguments{
\item{x}{A \code{processing_step}object}

\item{...}{further arguments passed to or from other methods(not
currenctly used).}
}
\value{
A tibble representation of x
}
\description{
Tidy a \code{processing_step}
}
\examples{
library(tidySpectR)

do_something <- function(x, y, z){
     print("doing something")
     as.list(match.call())
}

fun_call <- do_something(1,2,3)
# Remove unwanted arguments
fun_call$x <- NULL
fun_call

stp <- processing_step(fun_call, "test", "test_01")
tidy(stp)
}
