% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_template.R
\name{tidy.processing_template}
\alias{tidy.processing_template}
\title{Tidy a \code{processing_template}}
\usage{
\method{tidy}{processing_template}(x, ...)
}
\arguments{
\item{x}{A \code{processing_step}object}

\item{...}{further arguments passed to or from other methods(not
currenctly used).}
}
\value{
A tibble representation of x
}
\description{
Tidy a \code{processing_template}
}
\examples{
library(tidySpectR)

# Creating a masking step
template <- processing_template() \%>\%
            new_step(mask, list(from = 5, to = Inf), type = "mask") \%>\%
            new_step(bucket_uniform, list(N = 10), type = "bucket") 

tidy(template)
}
