% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vnoise}
\alias{vnoise}
\title{Find the max bin value in the noise region}
\usage{
vnoise(bin, r)
}
\arguments{
\item{bin}{A matrix containing the noise region
Cols 0-2  are : bins, bin_start, bin_end, then come samples (col-wise)}

\item{R}{resolution}
}
\value{
A dbl, maximum bin value in the noise region
}
\description{
Find the max bin value in the noise region
}
\keyword{internal}
