% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{wide2long}
\alias{wide2long}
\title{Pivot_wider on collection$data}
\usage{
wide2long(x, ...)
}
\arguments{
\item{x}{wide data tibble}

\item{...}{not currently used}
}
\description{
Pivot_wider on collection$data
}
\keyword{internal}
