% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{add_spectrum}
\alias{add_spectrum}
\alias{add_spctrum}
\alias{add_spectrum.collection}
\title{Add a spectrum to a collection}
\usage{
add_spectrum(obj, ...)

\method{add_spectrum}{collection}(
  obj,
  values,
  limits,
  id = deparse(substitute(values)),
  label = NA,
  ...
)
}
\arguments{
\item{obj}{A \code{collection} object.}

\item{...}{further arguments passed to or from other methods(not
currenctly used).}

\item{values}{A vector of intensity values.}

\item{limits}{A vector of length two with the upper and lower
limits of the measurement space.}

\item{id}{A unique identifier for the sample.}

\item{label}{A label for the sample.}
}
\value{
An updated version of \code{obj}.
}
\description{
Add a spectrum to a collection
}
\details{
It is assumed that the binning is uniform and correspond
to point measurments.

\code{labels} should be used to provide meanigful labels to the samples, e.g.
'treated' and 'control'.
}
