% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.collection}
\alias{autoplot.collection}
\title{Plot a collection of spectra in their current processing state.}
\usage{
\method{autoplot}{collection}(object, type = "raw", offset_x = 0, offset_y = 0, ...)
}
\arguments{
\item{object}{A \code{collection} object.}

\item{type}{A single character string for the type of plot to
produce with values "raw", "average" or "label_average". See details.}

\item{offset_x}{x axis offset value (in the axis unit) for the spectra}

\item{offset_y}{y axis offset value (in the axis unit) for the spectra}

\item{...}{Not currently used.}
}
\value{
a \code{ggplot} object
}
\description{
Plot a collection of spectra in their current processing state.
}
\details{
A \code{raw} plot displays individual spectra, combine it with offset values
to separate individual spectra along the x and/or y axes.

An \code{average} plot will represent only the average spectra and the range of
values within each bin. Offset values will be ignored.

A \code{label_average} plot will display the label-average spectrum and range of values
within each label. Use offset values to separate the spectra.

Further plot customization is possible (although limited) by passing ggplot
methods to the returned object
}
