% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{average_spectrum}
\alias{average_spectrum}
\alias{average_spectrum.collection}
\title{Calculate the average spectrum}
\usage{
average_spectrum(obj, group = "all", ...)

average_spectrum(obj, group = "all", ...)
}
\arguments{
\item{obj}{A\code{collection} object}

\item{group}{How to perform grouping. Either \code{all}or \code{labels}}

\item{...}{further arguments passed to or from other methods(not
currenctly used).}
}
\value{
An updated version of \code{collection}.
}
\description{
Averages spectra
}
