% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket_fromBreaks.R
\name{bin_sum}
\alias{bin_sum}
\title{Sums values over a single bin.
For internal use only.}
\usage{
bin_sum(lower, higher, data)
}
\description{
Sums values over a single bin.
For internal use only.
}
