% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket_aibin.R
\name{bin_value}
\alias{bin_value}
\title{Calculate bin value (Vb)}
\usage{
bin_value(dat, R)
}
\arguments{
\item{dat}{a date frame or tibble with the columns:
\describe{
\item{id}{unique identifier}
\item{bin_start}{left border of the bin}
\item{bin_end}{right border of the bin}
\item{values}{bin intensity}
}}

\item{R}{resolution}
}
\value{
A dbl, the bin value
}
\description{
Computes the inner product for each sample,
then the bin value over all samples.
}
\keyword{internal}
