% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket_aibin.R
\name{bucket_aibin}
\alias{bucket_aibin}
\alias{bucket_aibin.collection}
\title{Applies Adaptive Inteligent Binning on a collection}
\usage{
bucket_aibin(x, ...)

\method{bucket_aibin}{collection}(x, R, noise_region, ...)
}
\arguments{
\item{x}{A\code{collection} object to bucket}

\item{...}{further arguments passed to or from other methods(not
currenctly used).}

\item{R}{resolution value, strictly positive and typically in the interval
0 > R >= 1.}

\item{noise_region}{A\code{collection} object containing a noise_region}
}
\value{
An updated version of x
}
\description{
Adaptive Inteligent Binning (aibin) recursively finds bin edges
in existing bins by maximizing the information contained in each
newly created bin.
}
\details{
AI binning will first consider the entire spectra as a single bin.
It will then consider every position in the spectra as a potential
splitting candidate and calulate a quality value for each splitting candidate.
If the best splitting candidates fulfills quality requirements, the bin
will be split and the process will be repeated on the newly created bins.

In order to limit splitting in local minima of noisy regions a minimal quality
value is computed from the user-provided noise-region.
}
\note{
The amount of created bins is not directly dpeending of a high or low R.
If a maximal splitting is looked for, you should try several different values
in the possible rangeof R (see references). However the differences in binning should
limited to bins that are borderline noise and most likely has limited effect in
the downstream analysis.
}
\examples{
library(tidySpectR)

# First we normalize the dataset and extract noise and sepctra regions
normalized <- fa_nmr \%>\% 
                  normalize_internalStandard(3.58, 3.64)
spectra <- normalized \%>\% extract(0, 7.2)
noise <- normalized \%>\% mask(0, 7.2, overlaps = 'remove')

bucketted <- bucket_aibin(spectra, 0.2, noise)

}
\references{
De Meyer T, Sinnaeve D, Van Gasse B, Tsiporkova E, Rietzschel ER,
De Buyzere ML, Gillebert TC, Bekaert S, Martins JC, Van Criekinge W.
NMR-based characterization of metabolic alterations in hypertension
using an adaptive, intelligent binning algorithm. Anal Chem. 2008
May 15;80(10):3783-90. doi: 10.1021/ac7025964.
\url{https://pubmed.ncbi.nlm.nih.gov/18419139/}
}
