% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket_fromBreaks.R
\name{bucket_from_breaks}
\alias{bucket_from_breaks}
\alias{bucket_from_breaks.collection}
\title{Bucket a collection form a list of breaks}
\usage{
bucket_from_breaks(...)

\method{bucket_from_breaks}{collection}(obj, breaks, ...)
}
\arguments{
\item{...}{further arguments passed to or from other methods(not
currenctly used).}

\item{obj}{A\code{collection} object}

\item{breaks}{A vector of breaking values}
}
\value{
An updated version of \code{collection}.
}
\description{
Creates new bins delimited by the input breaks and sums up the intensity
values within each bin.
}
\details{
If omitted, the spectra limits will be added to the breaks.

Buckets will include the bin overlapping on the lower limit and exclude the bin
overlapping on the higher limit. This can results in a slight left-shift of the spectra.
Use with care.
}
\examples{
library(tidySpectR)

breaks = c(10, 8, 7.5, 7, 5, 0, -1)
bucket_from_breaks(fa_nmr, breaks)
}
