% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucketting.R
\name{bucket_optimized}
\alias{bucket_optimized}
\alias{bucket_optimized.collection}
\title{Performs optimized bucketting}
\usage{
bucket_optimized(...)

\method{bucket_optimized}{collection}(obj, initial_width, slackness, ...)
}
\arguments{
\item{...}{further arguments passed to or from other methods(not
currenctly used).}

\item{obj}{A\code{collection} object}

\item{initial_width}{Initial bin width to optimize}

\item{slackness}{fraction of initial width that defines
how far the boundary can move while searching for a local minimum}
}
\value{
An updated version of \code{collection}.
}
\description{
Bucket a collection of spectra using the Optimized
Bucketting Algorithm
}
\details{
The Optimized Bucketting Algorithm optimized bucket size by
searching local minima in the average spectrum.
}
\references{
Sousa SAA, Magalhaes A, Ferreira MMC. Optimized bucketing for NMR spectra:
Three case studies. Chemometr Intell Lab. 2013;122:93–102.
https://doi.org/10.1016/j.chemolab.2013.01.006
\url{https://doi.org/10.1016/j.chemolab.2013.01.006}
}
