% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket_uniform.R
\name{bucket_uniform}
\alias{bucket_uniform}
\alias{bucket_uniform.collection}
\title{Performs uniform bucketting}
\usage{
bucket_uniform(...)

\method{bucket_uniform}{collection}(obj, width = NULL, N = NULL, ...)
}
\arguments{
\item{...}{further arguments passed to or from other methods(not
currenctly used).}

\item{obj}{A\code{collection} object}

\item{width}{Bin width}

\item{N}{Number of bins to create}
}
\value{
An updated version of \code{collection}.
}
\description{
Bucket a collection of spectra to a given bin width or
number of bins
}
\details{
Provide either \code{width} or \code{N}
}
\examples{
library(tidySpectR)

bucket_uniform(fa_nmr, width= 0.04)

bucket_uniform(fa_nmr, N = 1500)
}
