% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{collection}
\alias{collection}
\alias{collection.default}
\alias{tidy.collection}
\title{Creates a spectra collection}
\usage{
collection(...)

\method{collection}{default}(...)

\method{tidy}{collection}(x, ...)
}
\arguments{
\item{...}{further arguments passed to or from other methods(not
currenctly used).}

\item{x}{A \code{collection} object.}
}
\value{
An object of class \code{collection} with sub-objects:
\item{data}{A tibble that contains the spectral data}
\item{normalized}{Information about the normalization of the dataset}
\item{bucketted}{Information about the bucketting of the dataset}
\item{labels}{Information about the labelling of the dataset}

A tibble with a variable number columns
depending on the binning of the data:
\item{id}{Unique sample identifier}
\item{label}{Label for the sample}
\item{bins}{Bin centers}
}
\description{
A collection is a matrix of spectral data and associated labels.
}
\details{
The collection is empty on creation, use \code{add_spectrum} to
populate it. It is recommended to populate a collection using \code{purrr::walk}
on a list of files (see \code{add_spectrum}).
}
\examples{
library(tidySpectR)

tidy(fa_nmr)
}
