% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket_aibin.R
\name{divide_bin}
\alias{divide_bin}
\title{Find a new division splitting position for the current bin}
\usage{
divide_bin(current_bin, R, vnoise)
}
\arguments{
\item{current_bin}{A \code{collection} object containing the current bin}

\item{R}{resolution}

\item{vnoise}{Maximal bin value for the noise region.}
}
\value{
A dbl, the new optimal splitting position or FALSE if not splitting is possible
}
\description{
Find a new division splitting position for the current bin
}
\keyword{internal}
