% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket_optimized.R
\name{find_local_min}
\alias{find_local_min}
\title{Find local minimum between two boundaries
For internal use only.}
\usage{
find_local_min(i, data, N, s)
}
\description{
Find local minimum between two boundaries
For internal use only.
}
