% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{mask}
\alias{mask}
\alias{mask.collection}
\title{Masks a spectral region}
\usage{
mask(...)

\method{mask}{collection}(obj, from, to, ...)
}
\arguments{
\item{...}{further arguments passed to or from other methods(not
currenctly used).}

\item{obj}{A\code{collection} object}

\item{from, to}{Coordinates of the region to mask.
Use -Inf or Inf to shorten the spectra.}
}
\value{
An updated version of \code{collection}.
}
\description{
\code{mask} discards a spectral region
}
\details{
The upper limit in not included: [min, max).
Bins that overlap on the limit will be removed.
}
\examples{
sample <- nmr_spectra \%>\% filter(ID == "20198124123")

coll <- collection() \%>\%
     add_spectrum(sample$values, 
                  c(min(sample$bins), max(sample$bins)),
                  id = "20198124123")

coll \%>\% mask(-2, 7.2)

# The order is not important
coll \%>\% mask(7.2, -2)

# Cropping edges
cropped <- coll \%>\%
             mask(-Inf, 0) \%>\%
             mask(Inf, 7.5)

}
