% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_probQuotient.R
\name{normalize_probQuotient}
\alias{normalize_probQuotient}
\alias{normalize_ProbQuotient.collection}
\alias{normalize_probQuotient.collection}
\title{Probabilistic Quotient normalization}
\usage{
normalize_probQuotient(x, ...)

\method{normalize_probQuotient}{collection}(x, ref = NULL, ...)
}
\arguments{
\item{x}{A\code{collection} object}

\item{...}{further arguments passed to or from other methods(not
currenctly used).}

\item{ref}{A reference spectra. Will use the median spectrum of the
collection if set to \code{NULL}.}
}
\value{
An updated version of \code{collection}.
}
\description{
Probabilistic Quoteint Normalization estimates for each spectra the
the most likely quotient derived form the ratio of the signal distribution
and that of a reference spectrum.
}
\examples{
library(tidySpectR)

normalize_probQuotient(fa_nmr)

# Provifding a reference spectrum
med <- median_spectrum(fa_nmr)
normalize_probQuotient(fa_nmr, ref = med)

}
\references{
Dieterle, F., Ross, A., Schlotterbeck, G., & Senn, H. (2006). Probabilistic
Quotient Normalization as Robust Method to Account for Dilution of Complex
Biological Mixtures. Application in1H NMR Metabonomics. Analytical Chemistry,
78(13), 4281–4290. doi:10.1021/ac051632c
\url{https://pubmed.ncbi.nlm.nih.gov/16808434/}
}
