% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_totalSpectrum.R
\name{normalize_totalSpectrum}
\alias{normalize_totalSpectrum}
\alias{normalize_totalSpectrum.collection}
\title{Normalizes spectra to their integrals}
\usage{
normalize_totalSpectrum(x, ...)

\method{normalize_totalSpectrum}{collection}(x, skip = FALSE, ...)
}
\arguments{
\item{x}{A\code{collection} object}

\item{...}{further arguments passed to or from other methods(not
currenctly used).}

\item{skip}{Skip the creation of of processor step. If TRUE, this step will not be added to
the list of processing steps. Typically reserved for nested function calls.}
}
\value{
An updated version of \code{collection}.
}
\description{
Sums up the signal over the whole spectrum and use it as a
normalization factor. This assumes that the total signal
varies proportionally to the sample concentration
}
\examples{
library(tidySpectR)

normalize_totalSpectrum(fa_nmr)
}
