% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{print.collection}
\alias{print.collection}
\title{Print a spectra collection}
\usage{
\method{print}{collection}(x, ...)
}
\arguments{
\item{x}{A \code{collection} object.}

\item{...}{further arguments passed to or from other methods(not
currenctly used).}
}
\value{
The original object (invisibly)
}
\description{
Print a spectra collection
}
\examples{
library(tidySpectR)
print(fa_nmr)
}
