% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_step.R, R/processing_template.R
\name{process.processing_step}
\alias{process.processing_step}
\alias{process}
\alias{process.processing_template}
\title{Applies a \code{processing_step} on a \code{collection} object}
\usage{
\method{process}{processing_step}(x, collection, ...)

process(...)

\method{process}{processing_template}(x, collection, verbose = FALSE, ...)
}
\arguments{
\item{x}{A \code{processing_template} object}

\item{collection}{a \code{collection} to process}

\item{...}{further arguments passed to or from other methods(not
currenctly used).}

\item{verbose}{Set to TRUE to show progress}
}
\description{
Applies a \code{processing_step} on a \code{collection} object

Process a \code{collection} object

Applies a \code{processing_template} on a \code{collection} object
}
\examples{
library(tidySpectR)

# Creating a masking step
stp <- 
   processing_step(
     fun = mask, 
     arglist = list(from = 5, to = Inf),
     type = "masking", 
     name = "masking")

process(stp, fa_nmr)
library(tidySpectR)

template <- processing_template() \%>\%
            new_step(mask, list(from = 5, to = Inf), type = "mask") \%>\%
            new_step(bucket_uniform, list(N = 10), type = "bucket") 
process(template, fa_nmr)
}
