% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_step.R
\name{processing_step}
\alias{processing_step}
\alias{processing_step.default}
\title{Stores specifications of a \code{collection} processing step}
\usage{
processing_step(...)

\method{processing_step}{default}(fun_call, type = "none", id = rand_id("none"), ...)
}
\arguments{
\item{...}{further arguments passed to or from other methods(not
currenctly used).}

\item{fun_call}{A function call created with as.list(match.call())}

\item{type}{A string description of the step}

\item{id}{A unique id associated with the step}
}
\value{
An object of calss \code{processing_step}
}
\description{
A \code{processing_step} stores the function call and associated
arguments to reproduce a processing step at a later time
point a a different collection object.
}
\examples{
library(tidySpectR)

do_something <- function(x, y, z){
     print("doing something")
     as.list(match.call())
}

fun_call <- do_something(1,2,3)
# Remove unwanted arguments
fun_call$x <- NULL
fun_call

processing_step(fun_call, "test", "test_01")
}
