% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pull_ids}
\alias{pull_ids}
\alias{pull_ids.collection}
\title{Pull the ids of a spectra collection}
\usage{
pull_ids(x, ...)

\method{pull_ids}{collection}(x, ...)
}
\arguments{
\item{x}{A spectra collection}

\item{...}{further arguments passed to or from other methods(not
currenctly used).}
}
\value{
A vector of chr.
}
\description{
For a collection, retrieve the sample identifiers
}
\examples{
library(tidySpectR)
pull_ids(fa_nmr)
}
