% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pull_numbin}
\alias{pull_numbin}
\alias{pull_numbin.collection}
\title{Pull the bin (or bucket) number of a spectra collection}
\usage{
pull_numbin(x, ...)

\method{pull_numbin}{collection}(x, ...)
}
\arguments{
\item{x}{A spectra collection}

\item{...}{further arguments passed to or from other methods(not
currenctly used).}
}
\value{
A tibble.
}
\description{
For a collection, retrieve the number of bins.
}
\examples{
library(tidySpectR)
pull_numbin(fa_nmr)
}
