% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_s_vast.R, R/step_vast.R, R/step_x_vast.R
\name{tidy.step_svast}
\alias{tidy.step_svast}
\alias{step_vast}
\alias{tidy.step_vast}
\alias{tidy.step_xvast}
\title{Applies VAST scaling on numeric data}
\usage{
\method{tidy}{step_svast}(x, ...)

step_vast(
  recipe,
  ...,
  scaling = "autoscale",
  role = NA,
  trained = FALSE,
  means = NULL,
  sds = NULL,
  cvs = NULL,
  na_rm = TRUE,
  skip = FALSE,
  id = rand_id("vast")
)

\method{tidy}{step_vast}(x, ...)

\method{tidy}{step_xvast}(x, ...)
}
\arguments{
\item{x}{A \code{step_vast} object.}

\item{...}{One or more selector functions to choose which
variables are affected by the step. See \code{\link[=selections]{selections()}}
for more details. For the \code{tidy} method, these are not
currently used.}

\item{scaling}{Either \code{autoscale} or \code{pareto}. Controls the scaling method.
See notes below.}

\item{role}{Not used by this step since no new variables are
created.}

\item{means}{A named numeric vector of means. This
is \code{NULL} until computed by \code{\link[=prep.recipe]{prep.recipe()}}.}

\item{sds}{A named numeric vector of stadard deviations. This
is \code{NULL} until computed by \code{\link[=prep.recipe]{prep.recipe()}}.}

\item{cvs}{A named numeric vector of variation coeficients. This
is \code{NULL} until computed by \code{\link[=prep.recipe]{prep.recipe()}}.}

\item{na_rm}{A logical value indicating whether \code{NA}
values should be removed when computing the standard deviation and mean.}
}
\value{
An updated version of \code{recipe} with the new step
added to the sequence of existing steps (if any). For the
\code{tidy} method, a tibble with columns \code{terms} (the
selectors or variables selected), \code{value} (the
standard deviations and means), and \code{statistic} for the type of value.
}
\description{
\code{step_vast} creates a \emph{specification} of a recipe
step that will perform VAST scaling on the columns
}
\details{
Variable Stability (VAST) scaling preforms centering and scaling followed
by a weighting of each variable by its variation coeficient.

The argument \code{scaling} controls which scaling method should be used before
variable weighting. \code{autoscale} will perform mean-centering and standard deviation
scaling while \code{pareto} will scale by the square-root of the standard deviation.
}
\examples{
# requires the recipes package
autoscale_vast <- 
  recipe(Species ~. , iris) \%>\%
  step_vast(all_predictors(), scaling = 'autoscale')

pareto_vast <- 
  recipe(Species ~. , iris) \%>\%
  step_vast(all_predictors(), scaling = 'pareto')
}
\references{
Keun H. C., Ebbels T. M. D., Antti H., Bollard M. E., Beckonert O., Holmes E.,
et al. (2003). Improved analysis of multivariate data by variable stability
scaling: application to NMR-based metabolic profiling. Anal. Chim. Acta 490,
265–276 10.1016/S0003-2670(03)00094-1
\url{https://www.sciencedirect.com/science/article/abs/pii/S0003267003000941}
}
\concept{normalization_methods}
\concept{preprocessing}
\keyword{datagen}
