% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket_aibin.R
\name{sub_bins_values}
\alias{sub_bins_values}
\title{Calculate the bin values for each splits of the current bin}
\usage{
sub_bins_values(x, i, current_bin, R)
}
\arguments{
\item{x}{spliting point}

\item{i}{splitting point index}

\item{current_bin}{A matrix containing the current bin}

\item{R}{resolution}
}
\value{
A Matrix with bin values:
\describe{
\item{candidate}{position of the new splitting point}
\item{vb1}{left-bin value}
\item{vb2}{right-bin value}
\item{vbsum}{sum of the left and righ bin values}
}
}
\description{
Calculate the bin values for each splits of the current bin
}
\keyword{internal}
